requirements.txt
flask==2.3.3
python-dotenv==1.0.0
requests==2.31.0
pytube==15.0.0
youtubesearchpython==1.6.6

